# swagger.yaml
swagger: '2.0'
info:
  description: Employee datebase enquiry system
  version: 1.0.0
  title: Swagger Employee
  termsOfService: http://swagger.io/terms/
  contact:
    email: skp.pradhan1989@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost:9092

tags:
- name: employee
  description: Details about employee
schemes:
- http
paths:
  /employee/{empId}:
    get:
      tags:
      - employee
      summary: Find employee by ID
      description: Returns a single employee details
      operationId: getEmployeeById
      produces:
      - application/json
      parameters:
      - name: empId
        in: path
        description: ID of employee to return
        required: true
        type: integer
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/Employee'
        400:
          description: Invalid ID supplied
        404:
          description: Employee not found
  /addemployee:
    post:
      tags:
      - employee
      summary: Adds employee record in database
      description: Adds employee record in database
      operationId: addEmployee
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Employee Object
        required: true
        schema:
          $ref: '#/definitions/Employee'
      responses:
        200:
          description: successful operation
          schema:
            type: string
        default:
          description: successful operation

definitions:
  Employee:
    type: object
    required:
    - id
    properties:
      id:
        type: integer
      empFullName:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      contact:
        type: string
      address:
        type: string
        description: Usemployeeer address
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io